[misc] Move subject identifier utility functions in a separate file
