+++ modules/data-entry/src/main/frontend/src/questionnaire/Subject.jsx
@@ -29,6 +29,7 @@ import { QUESTION_TYPES, SECTION_TYPES, ENTRY_TYPES } from "./FormEntry.jsx";
 import { usePageNameWriterContext } from "../themePage/Page.jsx";
 import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
 import MaterialTable from 'material-table';
+import { getSubjectIdFromPath, getHierarchyAsList, getTextHierarchy } from "./SubjectIdentifier";
 
 import {
   Avatar,
@@ -63,62 +64,6 @@ let createQueryURL = (query, type) => {
   url.searchParams.set("query", `SELECT * FROM [${type}] as n` + query);
   url.searchParams.set("limit", 1000);
   return url;
-}
-
-let defaultCreator = (node) => {
-  return {to: "/content.html" + node["@path"]}
-}
-
-// Extract the subject id from the subject path
-// returns null if the parameter is not a valid subject path (expected format: Subjects/<id>)
-export function getSubjectIdFromPath (path) {
-  return /Subjects\/([^.]+)/.exec(path || '')?.[1];
-}
-
-// Recursive function to get a flat list of parents
-export function getHierarchy (node, RenderComponent, propsCreator) {
-  let HComponent = RenderComponent || Link;
-  let hpropsCreator = propsCreator || defaultCreator;
-  let props = hpropsCreator(node);
-  let output = <React.Fragment>{node.type.label} <HComponent {...props}>{node.identifier}</HComponent></React.Fragment>;
-  if (node["parents"] && node["parents"].type) {
-    let ancestors = getHierarchy(node["parents"], HComponent, propsCreator);
-    return <React.Fragment>{ancestors} / {output}</React.Fragment>
-  } else {
-    return output;
-  }
-}
-
-// Recursive function to get a flat list of parents with no links and subject labels
-export function getTextHierarchy (node, withType = false) {
-  let type = withType ? (node?.["type"]?.["@name"] + " "): "";
-  let output = node.identifier;
-  if (node["parents"]) {
-    let ancestors = getTextHierarchy(node["parents"], withType);
-    return `${ancestors} / ${type}${output}`;
-  } else {
-    return type + output;
-  }
-}
-
-// Recursive function to get the list of ancestors as an array
-export function getHierarchyAsList (node, includeHomepage) {
-  let props = defaultCreator(node);
-  let parent = <>{node.type.label} <Link {...props} underline="hover">{node.identifier}</Link></>;
-  if (node["parents"]) {
-    let ancestors = getHierarchyAsList(node["parents"]);
-    ancestors.push(parent);
-    return ancestors;
-  } else {
-    let result = [parent];
-    includeHomepage && result.unshift(getHomepageLink(node));
-    return result;
-  }
-}
-
-export function getHomepageLink (subjectNode) {
-  let props = defaultCreator({"@path": `/Subjects#subjects:activeTab=${subjectNode?.type?.["@name"]}`});
-  return (<Link {...props} underline="hover">{subjectNode?.type?.subjectListLabel || "Subjects"}</Link>);
 }
 
 /**
@@ -365,7 +310,7 @@ function SubjectHeader(props) {
                />
             </div>
   );
-  let parentDetails = (subject?.data?.['parents'] && getHierarchyAsList(subject.data['parents'], true) || [getHomepageLink(subject?.data)]);
+  let parentDetails = (subject?.data?.['parents'] && getHierarchyAsList(subject.data['parents'], true));
 
   return (
     subject?.data &&
